Add instructions `lxi` and `mvi`
`mvi` copies the immediate value specified into the byte of memory
specified. `lxi` copies an immediate value into a register pair.

imm() required changes to add these instructions. Now, imm() is passed an enum
that specifies the size of the operands. It is to be IMM8 for 8 bits in all
cases barring the `lxi` instruction that requires IMM16.