Use accurate argument in imm()
Support for `lxi` and `mvi` required another conditional statement to
determine which argument to use -- `arg1` or `arg2`. The applicable
argument was set to a local variable `arg` in imm(). However, this local
variable did not replace the globals within the function were necessary
in previous commits.

Also, the last conditional within the else clause of the conditional
testing the first character of `arg` should be moved one scope upward
because each instruction must be processed during the second pass.